requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.27.0
pyyaml>=6.0
python-dotenv>=1.0.0
jinja2>=3.1.0
duckduckgo-search>=4.0
beautifulsoup4>=4.12.0
pytest>=8.0.0
